% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_test_analysis.R
\name{AggConfIntervals}
\alias{AggConfIntervals}
\title{GeoLift Confidence Intervals}
\usage{
AggConfIntervals(
  summary_augsynth,
  treatment_start_time,
  treatment_end_time,
  alpha
)
}
\arguments{
\item{summary_augsynth}{Result of calling summary function on augsynth object.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{alpha}{Significance level. Set to 0.1 by default.}
}
\value{
List that contains values for the ATT estimate, and the aggregate
Confidence Intervals.
}
\description{
Calculate ATT Aggregate Confidence Intervals w/conformal approach
}
