% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{ATT.loc.plot}
\alias{ATT.loc.plot}
\title{ATT plot function for GeoLift for each test location.}
\usage{
ATT.loc.plot(GeoLift, conf.level = 0.9, ...)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{conf.level}{Confidence level. By defaul 0.9.}

\item{...}{additional arguments}
}
\value{
ATT plot for GeoLift for each test location.
}
\description{
ATT plot function for GeoLift for each test location.
}
