% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{ATT.plot}
\alias{ATT.plot}
\title{Average ATT plot function for GeoLift.}
\usage{
ATT.plot(GeoLift, outcome = "Units", conf.level = 0.9, ...)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{outcome}{Name of the outcome variable. By default "Units".}

\item{conf.level}{Confidence level. By defaul 0.9.}
}
\value{
ATT plot for GeoLift.
}
\description{
Average ATT plot function GeoLift.
}
