% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{build_cluster}
\alias{build_cluster}
\title{Build cluster to parallelize operations across nodes in machine.}
\usage{
build_cluster(parallel_setup)
}
\arguments{
\item{parallel_setup}{A string indicating parallel workers set-up.
Can be "sequential" or "parallel".}
}
\value{
Cluster object that will parallelize operations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function builds the cluster and imports the necessary packages
to run Geolift: augsynth, dplyr and tidyr.
}
