% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_test_analysis.R
\name{ConfIntervals}
\alias{ConfIntervals}
\title{GeoLift Confidence Intervals}
\usage{
ConfIntervals(
  augsynth,
  treatment_start_time,
  treatment_end_time,
  alpha,
  stat_func,
  method,
  grid_size
)
}
\arguments{
\item{augsynth}{Augsynth object.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{alpha}{Significance level. Set to 0.1 by default.}

\item{stat_func}{Function to compute the test statistic.}

\item{method}{A string indicating the method used to calculate the
aggregate ATT Confidence Intervals.
\itemize{
\item{"conformal":}{ Conformal Inference.}
\item{"jackknife+":}{ Jackknife+.}
}}

\item{grid_size}{Number of grid points to use when inverting the hypothesis
test for Conformal Inference.}
}
\value{
List the Confidence Interval for the specified alpha.
}
\description{
Calculate the Aggregate ATT Confidence Intervals
}
