% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{CorrelationCoefficient}
\alias{CorrelationCoefficient}
\title{Calculate correlations between input markets.}
\usage{
CorrelationCoefficient(data, locs = c())
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locs}{List of markets to use in the calculation of the correlations.}
}
\value{
Correlation coefficient.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Calculate correlations between input markets
}
