% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cumulative_value.plot}
\alias{cumulative_value.plot}
\title{Plot the accumulated lift effect.}
\usage{
cumulative_value.plot(
  data,
  locations,
  treatment_start_time,
  treatment_end_time,
  location_id = "location",
  time_id = "time",
  Y_id = "Y",
  X = c(),
  alpha = 0.1,
  model = "none",
  fixed_effects = TRUE,
  method = "conformal",
  grid_size = 250,
  stat_test = "Total",
  conformal_type = "iid",
  ns = 1000,
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  title = "",
  subtitle = "",
  notes = "",
  ...
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{List of test locations.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{Y_id}{Name of the outcome variable (String).}

\item{X}{List of names of covariates.}

\item{alpha}{Significance level. Set to 0.1 by default.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
\item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
\item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
for smaller panels (less than 40 locations and 100 time-stamps.))}
\item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
to improve fit for larger panels (more than 40 locations and 100
time-stamps. }
\item{"best":}{ Fits the model with the lowest Scaled L2 Imbalance.}
}}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}

\item{method}{A string indicating the method used to calculate the
aggregate ATT Confidence Intervals.
\itemize{
\item{"conformal":}{ Conformal Inference. Defualt.}
\item{"jackknife+":}{ Jackknife+ (exclusively for stat_test = "Total").}
}}

\item{grid_size}{Number of grid points to use when inverting the hypothesis
test for Conformal Inference. Set to 250 by default.}

\item{stat_test}{A string indicating the test statistic.
\itemize{
\item{"Total":}{ The test statistic is the sum of all treatment effects, i.e. sum(abs(x)). Default.}
\item{"Negative":}{ One-sided test against positive effects i.e. -sum(x).
Recommended for Negative Lift tests.}
\item{"Positive":}{ One-sided test against negative effects i.e. sum(x).
Recommended for Positive Lift tests.}
}}

\item{conformal_type}{Type of conformal inference used. Can be either "iid" for Independent and identically
distributed or "block" for moving block permutations. Set to "iid" by default.}

\item{ns}{Number of resamples for "iid" permutations if \verb{conformal_type = "iid}. Set to 1000 by default.}

\item{treatment_end_date}{Character that represents a date in year-month-day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{title}{Character for the title of the plot. NULL by default.}

\item{subtitle}{Character for the subtitle of the plot. NULL by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{...}{additional arguments}
}
\value{
A ggplot object that shows the accumulated lift per time period.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Plot the accumulated lift effect.
}
