% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_data.R
\name{download_cluster_file}
\alias{download_cluster_file}
\title{Download Community Zones Clusters from Data For Good.}
\usage{
download_cluster_file(
  path_to_file_local,
 
    path_to_file_url = paste0("https://data.humdata.org/dataset/b7aaa3d7-cca2-4364-b7ce-afe3134194a2",
    "/resource/3c068b51-5f0d-4ead-80ba-97312ec034e4/download/",
    "data-for-good-at-meta-commuting-zones-march-2023.csv"),
  country_filter = NULL
)
}
\arguments{
\item{path_to_file_local}{Complete path where the downloaded file will be
stored in local.}

\item{path_to_file_url}{default url to look for Community Zones Clusters.}

\item{country_filter}{specific country in which clusters should be located.
Default is NULL.}
}
\value{
shapefile data.frame that holds values for region, country,
fbcz_id (cluster id), fbcz_id_num (cluster number), name, cz_gen_ds (date of
update), win_population, win_roads_km, area, geography (polygon with points
that form the cluster).
}
\description{
This method downloads the Community Zones Clusters CSV to local
and imports the file in a format compatible with the Shapefile library.
}
