% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{fn_treatment}
\alias{fn_treatment}
\title{Auxiliary function to generate the treatment variable D.}
\usage{
fn_treatment(df, locations, treatment_start_time, treatment_end_time)
}
\arguments{
\item{df}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{List of names of the test locations (String).}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}
}
\value{
Data frame with the additional treatment variable (D).
}
\description{
\code{fn_treatment} generates the treatment variable D, where
D = 1 for test locations in the test period and D = 0 otherwise.
}
