% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeoLift_PreTest}
\alias{GeoLift_PreTest}
\title{GeoLift Example Data - PreTest}
\format{
A data frame with historical information to run GeoLift power calculations:
\describe{
   \item{date}{Date in yyyy-mm-dd format}
   \item{location}{Name of the city}
   \item{Y}{Number of conversions}
}
}
\usage{
GeoLift_PreTest
}
\description{
Simulated data of daily conversions for 40 US cities for a period of 90 days.
This data provides historica pre-intervention information to run power
calculations for GeoLift.
}
\keyword{datasets}
