% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeoLift_Test}
\alias{GeoLift_Test}
\title{GeoLift Example Data - Test}
\format{
A data frame with historical information to run a GeoLift analysis:
\describe{
   \item{date}{Date in yyyy-mm-dd format}
   \item{location}{Name of the city}
   \item{Y}{Number of conversions}
}
}
\usage{
GeoLift_Test
}
\description{
Simulated data of daily conversions for 40 US cities for a period of 90
pre-intervention days followed by a 15-day campaign on the cities of chicago
and portland. This dataset provides information to run a GeoLift test for the
two markets that received an intervention.
}
\keyword{datasets}
