% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeoLift_Test_MultiCell}
\alias{GeoLift_Test_MultiCell}
\title{GeoLift Multi-Cell Example Data - Test}
\format{
A data frame with historical information to run a GeoLift analysis:
\describe{
\item{date}{Date in yyyy-mm-dd format}
\item{location}{Name of the city}
\item{Y}{Number of conversions}
}
}
\usage{
GeoLift_Test_MultiCell
}
\description{
Simulated Multi-Cell data of daily conversions for 40 US cities for a period of 90
pre-intervention days followed by a 15-day campaign on two cells: Cell 1 chicago
& cincinnati and Cell B honolulu and indianapolis.
This dataset provides information to run a Multi-Cell GeoLift test for the
two cells that received an intervention.
}
\keyword{datasets}
