% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{GeoLiftPower.search}
\alias{GeoLiftPower.search}
\title{Power calculations for unknown test market locations, number of
test markets, and test duration.}
\usage{
GeoLiftPower.search(
  data,
  treatment_periods,
  N = 1,
  horizon = -1,
  X = c(),
  Y_id = "Y",
  location_id = "location",
  time_id = "time",
  top_results = 5,
  alpha = 0.1,
  type = "pValue",
  normalize = FALSE,
  model = "none",
  fixed_effects = TRUE,
  stat_func = NULL,
  dtw = 0,
  ProgressBar = FALSE,
  run_stochastic_process = FALSE
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{treatment_periods}{List of treatment periods to calculate power for.}

\item{N}{List of number of test markets to calculate power for.}

\item{horizon}{An integer that defines at which time-stamp the power simulations
will start. This parameter allows the user to define which period is most relevant
for the test's current in-market dynamics (a very small horizon will include
simulations of time periods with dynamics that might not be relevant anymore).
Ideally the horizon should encompass at least of couple of times the test length.
For instance, for a power analysis with 180 days of historical data and a 15 day
test, we would recommend setting horizon to at least 150. By default horizon is set
to -1 which will  execute the smallest possible horizon with the provided data.}

\item{X}{List of names of covariates.}

\item{Y_id}{Name of the outcome variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{top_results}{Number of results to display.}

\item{alpha}{Significance Level. By default 0.1.}

\item{type}{Method of inference used in the analysis.
pValue=Provides conformal inference to provide the aggregate
p-value for the null hypothesis of no effect from the intervention.
The Default type is pValue.
Imbalance=Uses the model's Scaled L2 Imbalance metric.}

\item{normalize}{A logic flag indicating whether to scale the outcome which is
useful to accelerate computing speed when the magnitude of the data is large. The
default is FALSE.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
         \item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
         \item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
                          for smaller panels (less than 40 locations and 100 time-stamps.))}
         \item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
                         to improve fit for larger panels (more than 40 locations and 100
                         time-stamps. }
         }}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}

\item{stat_func}{Function to compute test statistic. NULL by default.}

\item{dtw}{Emphasis on Dynamic Time Warping (DTW), dtw = 1 focuses exclusively
on this metric while dtw = 0 (default) relies on correlations only.}

\item{ProgressBar}{A logic flag indicating whether to display a progress bar
to track progress. Set to FALSE by default.}

\item{run_stochastic_process}{A logic flag indicating whether to select test
markets through random sampling of the the similarity matrix. Given that
interpolation biases may be relevant if the synthetic control matches
the characteristics of the test unit by averaging away large discrepancies
between the characteristics of the test and the units in the synthetic controls,
it is recommended to only use random sampling after making sure all units are
similar. This parameter is set by default to FALSE.}
}
\value{
Data frame with the ordered list of best locations and their
average power.
}
\description{
\code{GeoLiftPower.search} provides power calculations for unknown
test markets, number of test locations, and test duration.
}
