% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{GeoPlot}
\alias{GeoPlot}
\title{Plotting function for Exploratory Analysis.}
\usage{
GeoPlot(
  data,
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  treatment_start = 0,
  KPI_id = "",
  notes = ""
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{treatment_start}{Treatment start (Default = 0).}

\item{KPI_id}{Outcome variable.}

\item{notes}{Additional notes.}
}
\value{
A plot of each location's time series.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{GeoPlot} takes a data frame used for GeoLift to generate
a plot of each location's time series.
}
