% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{GeoSampling}
\alias{GeoSampling}
\title{Multi-Cell Sampling Method.}
\usage{
GeoSampling(data, k = 2, sampling_method = "systematic")
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{k}{Number of partitions or cells. k = 2 by default.}

\item{sampling_method}{Sampling Method used to create the k partitions.
\itemize{
\item{"systematic":}{ Systematic sampling based on the KPI (Y).
Defualt.}
}}
}
\value{
A data frame designating the cell per location. The resulting object
contains two columns: location and cell.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{GeoSampling} performs a sampling methodology to split the
data set into k similar partitions.
}
