% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_correlation_coefficient}
\alias{get_correlation_coefficient}
\title{Calculate correlations between input markets.}
\usage{
get_correlation_coefficient(data, locs = c())
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locs}{List of markets to use in the calculation of the correlations.}
}
\value{
Correlation coefficient.
}
\description{
Calculate correlations between input markets
}
