% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_data.R
\name{get_location_lat_long}
\alias{get_location_lat_long}
\title{Match location units to spatial location.}
\usage{
get_location_lat_long(data, country_name)
}
\arguments{
\item{data}{dataset that holds locations that will be used.}

\item{country_name}{country name in english to make lat long match easier.}
}
\description{
Find location latitude and longitude using Gmaps API.
}
