% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{GetMultiCellWeights}
\alias{GetMultiCellWeights}
\title{Function to obtain the synthetic control weights.}
\usage{
GetMultiCellWeights(x, test_markets = list())
}
\arguments{
\item{x}{A \code{MultiCellMarketSelection} object.}

\item{test_markets}{List of market IDs per cell. The list must contain exactly
k numeric values corresponding to the power analysis. The recommended layout is
\code{list(cell_1 = 1, cell2 = 1, cell3 = 1,...)}.}
}
\value{
Data-frame with the locations and the synthetic control weights for each cell.
}
\description{
\code{GetMultiCellWeights} returns the synthetic control weights as a data frame
for a given test set-up.
}
