% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{GetWeights}
\alias{GetWeights}
\title{Function to obtain the synthetic control weights.}
\usage{
GetWeights(
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  X = c(),
  data,
  locations,
  pretreatment_end_time,
  model = "none",
  fixed_effects = TRUE
)
}
\arguments{
\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{X}{List of names of covariates.}

\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{List of test locations.}

\item{pretreatment_end_time}{Time index of the last pre-treatment period.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
\item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
\item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
for smaller panels (less than 40 locations and 100 time-stamps.))}
\item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
to improve fit for larger panels (more than 40 locations and 100
time-stamps. }
\item{"best":}{ Fits the model with the lowest Scaled L2 Imbalance.}
}}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}
}
\value{
Data frame with the locations and weights.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{GetWeights} returns the synthetic control weights as a data frame
for a given test set-up.
}
