% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{incremental.plot}
\alias{incremental.plot}
\title{Daily Incrementality plot function for GeoLift output.}
\usage{
incremental.plot(
  GeoLift,
  treatment_end_date = NULL,
  main_title = "Daily incremental value",
  subtitle = "GeoLift analysis",
  ...
)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{treatment_end_date}{treatment_end_date Character that represents a date in year-month-day format.}

\item{main_title}{Character for the title of the plot. "Daily incremental value" by default.}

\item{subtitle}{Character for the subtitle of the plot. "GeoLift analysis" by default.}

\item{...}{additional arguments}
}
\value{
Daily Incremental plot.
}
\description{
\code{incremental_plot} returns chart for daily incrementality using GeoLift output.
}
