% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{Lift.loc.plot}
\alias{Lift.loc.plot}
\title{Lift plot function for GeoLift for each test location.}
\usage{
Lift.loc.plot(GeoLift, outcome, ...)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{outcome}{Name of the outcome variable. By default "Units".}
}
\value{
Lift plot function for GeoLift for each test location.
}
\description{
Lift plot function for GeoLift for each test location.
}
