% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{Lift.plot}
\alias{Lift.plot}
\title{Aggregate Lift plot function for GeoLift.}
\usage{
Lift.plot(
  GeoLift,
  treatment_end_date = NULL,
  title = "Daily observations per test group",
  subtitle = "",
  ...
)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{title}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{...}{additional arguments}
}
\value{
Aggregate Lift plot.
}
\description{
Aggregate Lift plot function for GeoLift.
}
