% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{limit_test_markets}
\alias{limit_test_markets}
\title{Limit the amount of test market potential elements.}
\usage{
limit_test_markets(
  similarity_matrix,
  treatment_sizes,
  run_stochastic_process = FALSE
)
}
\arguments{
\item{similarity_matrix}{Matrix that sorts each location in terms
of descending correlation.}

\item{treatment_sizes}{Vector of number of test markets to calculate
power for.}

\item{run_stochastic_process}{A logic flag indicating whether to select test
markets through random sampling of the the similarity matrix. Given that
interpolation biases may be relevant if the synthetic control matches
the characteristics of the test unit by averaging away large discrepancies
between the characteristics of the test and the units in the synthetic controls,
it is recommended to only use random sampling after making sure all units are
similar. Set to FALSE by default.}
}
\value{
Returns the limited number of treatment sizes if required. If the condition
is not met, it returns the original treatment sizes.
}
\description{
\code{limit_test_markets} determines if the user would like
to find combinations of treatments that is higher than half
the total amount of locations available while running a random
assignment of locations. If this is the case, it will cap the
treatment sizes at half the total amount of locations.
}
