% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_data.R
\name{location_to_cluster_matching}
\alias{location_to_cluster_matching}
\title{Match all locations to the cluster they belong.}
\usage{
location_to_cluster_matching(
  location_points,
  clusters,
  longitude_col_name = "Longitude",
  latitude_col_name = "Latitude"
)
}
\description{
Match all locations to the cluster they belong.
}
