% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{MarketCorrelations}
\alias{MarketCorrelations}
\title{Auxiliary function to calculate correlations between input markets.}
\usage{
MarketCorrelations(
  data,
  location_id = "location",
  time_id = "time",
  Y_id = "Y",
  dtw = 0,
  markets = NULL
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{Y_id}{Name of the outcome variable (String).}

\item{dtw}{Emphasis on Dynamic Time Warping (DTW), dtw = 1 focuses exclusively
on this metric while dtw = 0 (default) relies on correlations only.}

\item{markets}{List of markets to use in the calculation of the correlations.}
}
\value{
MarketMatching object.
}
\description{
Auxiliary function to calculate correlations between input markets.
}
