% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_test_power.R
\name{MarketSelection}
\alias{MarketSelection}
\title{Market selection tool.}
\usage{
MarketSelection(
  data,
  location_id = "location",
  time_id = "time",
  Y_id = "Y",
  dtw = 0,
  exclude_markets = c()
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{Y_id}{Name of the outcome variable (String).}

\item{dtw}{Emphasis on Dynamic Time Warping (DTW), dtw = 1 focuses exclusively
on this metric while dtw = 0 (default) relies on correlations only.}

\item{exclude_markets}{A list of markets or locations that won't be considered
for the test market selection, but will remain in the pool of controls. Empty
list by default.}
}
\value{
Matrix of the best markets. The second to last columns show
the best to worst market matches for the location in the first
column.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{MarketSelection} helps calculate the best markets based
on Dynamic Time Warping between the locations' time-series.
}
