% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{MultiCellPower}
\alias{MultiCellPower}
\title{Multi-Cell Power Analysis Method.}
\usage{
MultiCellPower(
  x,
  test_markets = list(),
  effect_size = seq(-0.25, 0.25, 0.05),
  lookback_window = NULL,
  side_of_test = NULL,
  plot = TRUE,
  actual_values = TRUE,
  smoothed_values = FALSE,
  show_mde = TRUE,
  breaks_x_axis = 10
)
}
\arguments{
\item{x}{A \code{MultiCellMarketSelection} object.}

\item{test_markets}{List of market IDs per cell. The list must contain exactly
k numeric values corresponding to the power analysis. The recommended layout is
\code{list(cell_1 = 1, cell2 = 1, cell3 = 1,...)}.}

\item{effect_size}{A vector of effect sizes to test by default a
sequence between 0 - 25 percent in 5 percent increments: seq(0,0.25,0.05).
Only input sequences that are entirely positive or negative and that include
zero.}

\item{lookback_window}{A number indicating how far back in time the simulations
for the power analysis should go. For instance, a value equal to 5 will simulate
power for the last five possible tests. By default lookback_window = 1 which
will only execute the most recent test based on the data.}

\item{side_of_test}{A string indicating whether confidence will be determined
using a one sided or a two sided test. If set to NULL (default), the \code{side_of_test}
will be populated by the value specified in \code{MultiCellMarketSelection}.
\itemize{
\item{"two_sided":}{ The test statistic is the sum of all treatment effects, i.e. sum(abs(x)). Defualt.}
\item{"one_sided":}{ One-sided test against positive or negaative effects i.e.
If the effect being applied is negative, then defaults to -sum(x). H0: ES >= 0; HA: ES < 0.
If the effect being applied is positive, then defaults to sum(x). H0: ES <= 0; HA: ES > 0.}
}}

\item{plot}{Logic flag indicating whether to plot all power curves. Set to TRUE
by default.}

\item{actual_values}{Logic flag indicating whether to include in the plot
the actual values. TRUE by default.}

\item{smoothed_values}{Logic flag indicating whether to include in the plot
the smoothed values. TRUE by default.}

\item{show_mde}{Logic flag indicating whether to include in the plot
the positive and negative MDEs. FALSE by default.}

\item{breaks_x_axis}{Numeric value indicating the number of breaks in the
x-axis of the power plot. You may get slightly more or fewer breaks that
requested based on \code{breaks_pretty()}. Set to 10 by default.}
}
\value{
A 'MultiCellPower' object with three objects:
\itemize{
\item{"PowerCurves":}{ List containing the Power Curves for each cell.}
\item{"data":}{ The input data.}
\item{"test_details":}{ The test details.}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{MultiCellPower} performs a Power-Analysis for a given set of Market Selections
obtained from \code{MultiCellMarketSelection}.
}
