% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{NumberLocations}
\alias{NumberLocations}
\title{Power Calculation to determine the number of test periods
with unknown test locations.}
\usage{
NumberLocations(
  data,
  number_locations = c(),
  treatment_periods,
  n_sim = 50,
  X = c(),
  Y_id = "Y",
  location_id = "location",
  time_id = "time",
  plot = TRUE,
  power = 0.8,
  conf.level = 0.9
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{number_locations}{List of number of locations to test. If not specified,
the number of locations will be computed by percentiles up to half of the
total number of locations.}

\item{treatment_periods}{Number of treatment periods.}

\item{n_sim}{Number of simulations.}

\item{X}{List of covariate names.}

\item{Y_id}{Name of the outcome variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{plot}{Plots results when TRUE.}

\item{power}{Power level. By default 0.8.}

\item{conf.level}{Confidence level. By default 0.9.}
}
\value{
Table of average power by number of locations.
}
\description{
\code{NumberLocations} calculates power to determine the
number of test periods with unknown test locations.
}
