% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.BestPreTreatmentLength}
\alias{plot.BestPreTreatmentLength}
\title{Plot to observe Best Start Time Period}
\usage{
\method{plot}{BestPreTreatmentLength}(x, ...)
}
\arguments{
\item{x}{BestPreTreatmentLength object.}

\item{...}{additional arguments}
}
\value{
Best Pre Treatment Length plot.
}
\description{
Plot to observe Best Start Time Period
}
