% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.BestStartTimePeriod}
\alias{plot.BestStartTimePeriod}
\title{Plot to observe Best Start Time Period}
\usage{
\method{plot}{BestStartTimePeriod}(x, ...)
}
\arguments{
\item{x}{BestStartTimePeriod object.}

\item{...}{additional arguments}
}
\value{
Best Start Time Period plot.
}
\description{
Plot to observe Best Start Time Period
}
