% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{plot.GeoLift}
\alias{plot.GeoLift}
\title{Plot for GeoLift.}
\usage{
\method{plot}{GeoLift}(
  x,
  type = "Lift",
  treatment_end_date = NULL,
  title = "",
  subtitle = "",
  ...
)
}
\arguments{
\item{x}{GeoLift object.}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted. If type is set to "Incrementality",
daily incremental values are plotted.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{title}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{...}{additional arguments}
}
\value{
GeoLift plot.
}
\description{
Plot for GeoLift objects.
}
