% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.GeoLiftMarketSelection}
\alias{plot.GeoLiftMarketSelection}
\title{Plotting function for GeoLiftMarketSelection}
\usage{
\method{plot}{GeoLiftMarketSelection}(
  x,
  market_ID = 0,
  print_summary = TRUE,
  actual_values = TRUE,
  smoothed_values = TRUE,
  show_mde = FALSE,
  breaks_x_axis = 10,
  side_of_test = "two_sided",
  ...
)
}
\arguments{
\item{x}{A GeoLiftMarketSelection object.}

\item{market_ID}{Numeric value indicating the market to be plotted. This
value should reflect a valid ID from the BestMarkets data frame of the
\code{GeoLiftMarketSelection} output.}

\item{print_summary}{Logic flag indicating whether to print model metrics
from the latest possible test. Set to TRUE by default.}

\item{actual_values}{Logic flag indicating whether to include in the plot
the actual values. TRUE by default.}

\item{smoothed_values}{Logic flag indicating whether to include in the plot
the smoothed values. TRUE by default.}

\item{show_mde}{Logic flag indicating whether to include in the plot
the positive and negative MDEs. FALSE by default.}

\item{breaks_x_axis}{Numeric value indicating the number of breaks in the
x-axis of the power plot. You may get slightly more or fewer breaks that
requested based on \code{breaks_pretty()}. Set to 10 by default.}

\item{side_of_test}{A string indicating whether confidence will be determined
using a one sided or a two sided test.}

\item{...}{additional arguments}
}
\value{
GeoLiftMarketSelection plot.
}
\description{
Plotting function for \code{GeoLiftMarketSelection}. This function plots the
latest possible test given the data and duration as well as the power curve
across historical simulations.
}
