% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{plot.GeoLiftMultiCell}
\alias{plot.GeoLiftMultiCell}
\title{Plot for GeoLiftMultiCell}
\usage{
\method{plot}{GeoLiftMultiCell}(
  x,
  type = "Lift",
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  title = "",
  post_treatment_periods = 0,
  ...
)
}
\arguments{
\item{x}{GeoLiftMultiCell object.}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted. If type is set to "Incrementality",
daily incremental values are plotted.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{title}{String for the title of the plot. Empty by default.}

\item{post_treatment_periods}{Number of post-treatment periods. Zero by default.}

\item{...}{additional arguments}
}
\value{
GeoLiftMultiCell plots.
}
\description{
Plot for GeoLiftMultiCell objects.
}
