% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.GeoLiftPower}
\alias{plot.GeoLiftPower}
\title{Plotting function for GeoLiftPower.}
\usage{
\method{plot}{GeoLiftPower}(
  x,
  power = 0.8,
  return_power_table = FALSE,
  actual_values = TRUE,
  smoothed_values = TRUE,
  show_mde = FALSE,
  ...
)
}
\arguments{
\item{x}{GeoLiftPower object.}

\item{power}{Power level. By default 0.8.}

\item{return_power_table}{Plot the table of power estimates. FALSE by default.}

\item{actual_values}{Logic flag indicating whether to include in the plot
the actual values. TRUE by default.}

\item{smoothed_values}{Logic flag indicating whether to include in the plot
the smoothed values. TRUE by default.}

\item{show_mde}{Logic flag indicating whether to include in the plot
the positive and negative MDEs. FALSE by default.}

\item{...}{additional arguments}
}
\value{
GeoLiftPower plot.
}
\description{
Plotting function for \code{GeoLiftPower}. The function smooths the power curve
for ease of interpretation.
}
