% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{plot.MultiCellMarketSelection}
\alias{plot.MultiCellMarketSelection}
\title{Plotting function for MultiCellMarketSelection}
\usage{
\method{plot}{MultiCellMarketSelection}(
  x,
  cell_ID = 0,
  market_ID = 0,
  print_summary = TRUE,
  actual_values = TRUE,
  smoothed_values = TRUE,
  show_mde = FALSE,
  breaks_x_axis = 10,
  ...
)
}
\arguments{
\item{x}{A MultiCellMarketSelection object.}

\item{cell_ID}{Numeric value indicating the cell.}

\item{market_ID}{Numeric value indicating the market to be plotted. This
value should reflect a valid ID from the \code{MultiCellMarketSelection} output.}

\item{print_summary}{Logic flag indicating whether to print model metrics
from the latest possible test. Set to TRUE by default.}

\item{actual_values}{Logic flag indicating whether to include in the plot
the actual values. TRUE by default.}

\item{smoothed_values}{Logic flag indicating whether to include in the plot
the smoothed values. TRUE by default.}

\item{show_mde}{Logic flag indicating whether to include in the plot
the positive and negative MDEs. FALSE by default.}

\item{breaks_x_axis}{Numeric value indicating the number of breaks in the
x-axis of the power plot. You may get slightly more or fewer breaks that
requested based on \code{breaks_pretty()}. Set to 10 by default.}

\item{...}{additional arguments}
}
\value{
MultiCellMarketSelection plot.
}
\description{
Plotting function for \code{MultiCellMarketSelection}. This function plots the
latest possible test given the data and duration as well as the power curve
across historical simulations for a given merket_id and cell_id.
}
