% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_test_analysis.R
\name{print.GeoLift}
\alias{print.GeoLift}
\title{Print pretty GeoLift output.}
\usage{
\method{print}{GeoLift}(x, ...)
}
\arguments{
\item{x}{GeoLift object.}

\item{...}{Optional arguments}
}
\value{
GeoLift output message
}
\description{
Print GeoLift output.
}
