% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{print.GeoLiftMarketFinder}
\alias{print.GeoLiftMarketFinder}
\title{Printing method for GeoLiftMarketFinder.}
\usage{
\method{print}{GeoLiftMarketFinder}(x, ...)
}
\arguments{
\item{x}{GeoLiftMarketFinder object.}

\item{...}{Optional arguments.}
}
\value{
A data frame.
}
\description{
Printing method for \code{GeoLiftMarketFinder}. The function will print
the best markets when called.
}
