% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_test_power.R
\name{print.GeoLiftMarketSelection}
\alias{print.GeoLiftMarketSelection}
\title{Printing method for GeoLiftMarketSelection}
\usage{
\method{print}{GeoLiftMarketSelection}(x, ...)
}
\arguments{
\item{x}{GeoLiftMarketSelection object.}

\item{...}{Optional arguments.}
}
\value{
A data frame.
}
\description{
Printing method for \code{GeoLiftMarketSelection}. The function will print
the best markets when called.
}
