% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{print.GeoLiftMultiCell}
\alias{print.GeoLiftMultiCell}
\title{Print pretty GeoLiftMultiCell output.}
\usage{
\method{print}{GeoLiftMultiCell}(x, ...)
}
\arguments{
\item{x}{GeoLiftMultiCell object.}

\item{...}{Optional arguments}
}
\value{
GeoLiftMultiCell output message
}
\description{
Print GeoLiftMultiCell output.
}
