% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{print.MultiCellMarketSelection}
\alias{print.MultiCellMarketSelection}
\title{Print pretty MultiCellMarketSelection output.}
\usage{
\method{print}{MultiCellMarketSelection}(x, ...)
}
\arguments{
\item{x}{MultiCellMarketSelection object.}

\item{...}{Optional arguments}
}
\value{
MultiCellMarketSelection output message
}
\description{
Print MultiCellMarketSelection output.
}
