% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{print.MultiCellPower}
\alias{print.MultiCellPower}
\title{Print pretty MultiCellPower output.}
\usage{
\method{print}{MultiCellPower}(x, ...)
}
\arguments{
\item{x}{MultiCellPower object.}

\item{...}{Optional arguments}
}
\value{
MultiCellPower output message
}
\description{
Print MultiCellPower output.
}
