% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{print.MultiCellWinner}
\alias{print.MultiCellWinner}
\title{Print pretty MultiCellWinner output.}
\usage{
\method{print}{MultiCellWinner}(x, ...)
}
\arguments{
\item{x}{MultiCellWinner object.}

\item{...}{Optional arguments}
}
\value{
MultiCellWinner output message
}
\description{
Print MultiCellWinner output.
}
