% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{print.summary.GeoLift}
\alias{print.summary.GeoLift}
\title{Summary plotting method for GeoLift.}
\usage{
\method{print}{summary.GeoLift}(x, ...)
}
\arguments{
\item{x}{GeoLift summary object.}

\item{...}{Optional arguments.}
}
\value{
Plot GeoLift summary.
}
\description{
Summary plotting method for GeoLift.
}
