% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{pvalue}
\alias{pvalue}
\title{P-value calculation for GeoLift.}
\usage{
pvalue(augsyn)
}
\arguments{
\item{augsyn}{GeoLift object.}
}
\value{
P-value.
}
\description{
\code{pvalue} calculates the p-value for a GeoLift object.
}
