% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_data.R
\name{ReplaceTreatmentSplit}
\alias{ReplaceTreatmentSplit}
\title{Replace Treatment locations to be able to use them for continuous GeoLift studies.}
\usage{
ReplaceTreatmentSplit(
  treatment_locations,
  data,
  treatment_start_time,
  treatment_end_time,
  model,
  verbose = FALSE
)
}
\arguments{
\item{treatment_locations}{Vector of locations where the treatment was applied.}

\item{data}{DataFrame that GeoLfit will use to determine a result.
Should be the output of \code{GeoDataRead}.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
\item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
\item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
for smaller panels (less than 40 locations and 100 time-stamps.))}
\item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
to improve fit for larger panels (more than 40 locations and 100
time-stamps. }
}}

\item{verbose}{boolean that determines if processing messages will be shown.}
}
\value{
list that contains:
\itemize{
\item{"data":}{ Data with replaced values for treatment locations during treatment period.}
\item{"l2_imbalance_df":}{ Ranking of treatment locations based on L2 imbalance for each iteration.}
}
}
\description{
\code{ReplaceTreatmentSplit} chooses the best treatment location to replace with their
control, given the L2 imbalance that each individual treatment has. Then
re-estimates the remaining treatment locations using the replaced treatment as
part of the control donor pool.
}
