% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_test_power.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations for treatment markets.}
\usage{
run_simulations(
  data,
  treatment_combinations,
  treatment_durations,
  effect_sizes = 0,
  side_of_test = "two_sided",
  lookback_window = 1,
  parallel = TRUE,
  pb = NULL,
  cpic = 0,
  X = c(),
  normalize = FALSE,
  fixed_effects = TRUE,
  model = "none"
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{treatment_combinations}{A matrix of treatment locations.
Each row symbolizes a combination, each column one element of the combination.}

\item{treatment_durations}{Expected durations of the experiment.}

\item{effect_sizes}{A vector of effect sizes to simulate.}

\item{side_of_test}{A string indicating whether confidence will be determined
using a one sided or a two sided test.
\itemize{
\item{"two_sided":}{ The test statistic is the sum of all treatment effects, i.e. sum(abs(x)). Defualt.}
\item{"one_sided":}{ One-sided test against positive or negaative effects i.e.
If the effect being applied is negative, then defaults to -sum(x). H0: ES >= 0; HA: ES < 0.
If the effect being applied is positive, then defaults to sum(x). H0: ES <= 0; HA: ES > 0.}
}}

\item{lookback_window}{A number indicating how far back in time the simulations
for the power analysis should go. For instance, a value equal to 5 will simulate
power for the last five possible tests. By default lookback_window = 1 which
will only execute the most recent test based on the data.}

\item{parallel}{A logic flag indicating whether to use parallel computing to
speed up calculations. Set to TRUE by default.}

\item{pb}{ProgressBar object. NULL by default.}

\item{cpic}{Number indicating the Cost Per Incremental Conversion.}

\item{X}{List of names of covariates.}

\item{normalize}{A logic flag indicating whether to scale the outcome which is
useful to accelerate computing speed when the magnitude of the data is large. The
default is FALSE.}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
\item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
\item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
for smaller panels (less than 40 locations and 100 time-stamps.))}
\item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
to improve fit for larger panels (more than 40 locations and 100
time-stamps. }
}}
}
\value{
DataFrame that contains:
\itemize{
\item{"location":}{ Test locations.}
\item{"pvalue":}{ P Value.}
\item{"tp":}{ Time period index.}
\item{"es":}{ Effect Size used for the simulation.}
\item{"treatment_start_time":}{ Treatment start time for the simulation}
\item{"investment":}{ Estimated Investment}
\item{"ScaledL2Imbalance":}{ Scaled L2 Imbalance metric}
\item{"att_estimator":}{ Detected Average Treatment on the Treated}
\item{"detected_lift":}{ Detected \% Lift by GeoLift for an effect size}
}
}
\description{
\code{run_simulations} computes simulations for each treatment market.
}
