% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{summary.GeoLift}
\alias{summary.GeoLift}
\title{Summary method for GeoLift.}
\usage{
\method{summary}{GeoLift}(GeoLift)
}
\arguments{
\item{GeoLift}{GeoLift object.}
}
\value{

}
\description{
GeoLift summary output with additional information about the
test.
}
