% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{TrimControls}
\alias{TrimControls}
\title{Helper to trim controls for shorter run-times with minimal loss
of precision.}
\usage{
TrimControls(
  data,
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  max_controls = 20,
  test_locations = c(),
  forced_control_locations = c()
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{max_controls}{Max number of controls, recommended 4x-5x
the number of test locations.}

\item{test_locations}{List of test locations.}

\item{forced_control_locations}{List of locations to be forced
as controls.}
}
\value{
A data frame with reduced control locations.
}
\description{
As the number of controls for a Geo test increases, the model
complexity grows as does the algorithm's run-time. However,
there are diminishing marginal returns in adding too many control
locations, especially if their time-series are very similar.
\code{TrimControls} provides a method to trim the number of controls
in order to reduce run-times with minimal loss of precision. In general,
it is recommended to have 4 to 5 times the number of controls locations
than the ones we have for test locations.
}
