% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{type_of_test}
\alias{type_of_test}
\title{Decides type of statistical function being applied for Conformal
Inference.}
\usage{
type_of_test(side_of_test = "two_sided", alternative_hypothesis = NULL)
}
\arguments{
\item{side_of_test}{A string indicating whether confidence will be determined
using a one sided or a two sided test.
\itemize{
         \item{"two_sided":}{ The test statistic is the sum of all treatment effects, i.e. sum(abs(x)). Defualt.}
         \item{"one_sided":}{ One-sided test against positive or negaative effects i.e. 
         If the effect being applied is negative, then defaults to -sum(x). H0: ES >= 0; HA: ES < 0.
         If the effect being applied is positive, then defaults to sum(x). H0: ES <= 0; HA: ES > 0.}
         }}

\item{alternative_hypothesis}{A string indicating what is the alternative hypothesis being tested. Defaults to NULL. 
\itemize{
         \item{"negative":}{ H0: ES >= 0; HA: ES < 0.}
         \item{"positive":}{ H0: ES <= 0; HA: ES > 0.}
}}
}
\value{
Statistical function being used to sum ATT effects over all treatment periods.
}
\description{
\code{type_of_test} returns stat_func being used for GeoLiftPower;
GeoLiftPowerFinder & GeoLift.
}
