% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_test_analysis.R
\name{ASCMExecution}
\alias{ASCMExecution}
\title{Augsynth execution.}
\usage{
ASCMExecution(
  data,
  treatment_locations,
  treatment_start_time,
  treatment_end_time,
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  X = c(),
  model = "none",
  fixed_effects = TRUE
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{treatment_locations}{Vector of treatment locations.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{X}{Vector with covariates names.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
\item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
\item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
for smaller panels (less than 40 locations and 100 time-stamps.))}
\item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
to improve fit for larger panels (more than 40 locations and 100
time-stamps. }
\item{"best:}{ Fits the model with the lowest Scaled L2 Imbalance.}
}}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}
}
\value{
List object that contains:
\itemize{
\item{"augsynth_model":}{ Generalized Augmented Sunthetic Controls results.}
\item{"data":}{ Properly formatted data to fit GeoLift.}
\item{"treatment_locations":}{ Treatment locations being used in lower case.}
}
}
\description{
\code{ASCMExecution} executes the augmented synthetic controls package.
}
