% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{ATT.plot}
\alias{ATT.plot}
\title{Average ATT plot function for GeoLift.}
\usage{
ATT.plot(GeoLift, outcome = "Units", main = "", subtitle = "", notes = "", ...)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{outcome}{Name of the outcome variable. By default "Units".}

\item{main}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{...}{additional arguments}
}
\value{
ATT plot for GeoLift.
}
\description{
Average ATT plot function GeoLift.
}
