% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{build_cluster}
\alias{build_cluster}
\title{Build cluster to parallelize operations across nodes in machine.}
\usage{
build_cluster(parallel_setup, import_augsynth_from)
}
\arguments{
\item{parallel_setup}{A string indicating parallel workers set-up. 
Can be "sequential" or "parallel".}

\item{import_augsynth_from}{Points to where the augsynth package
should be imported from to send to the nodes.}
}
\value{
Cluster object that will parallelize operations.
}
\description{
This function builds the cluster and imports the necessary packages
to run Geolift: augsynth, dplyr and tidyr.
}
