% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{cumulative_lift}
\alias{cumulative_lift}
\title{Calculate cumulative lift}
\usage{
cumulative_lift(
  data,
  treatment_locations,
  treatment_start_period,
  treatment_end_period,
  location_id = "location",
  time_id = "time",
  Y_id = "Y"
)
}
\arguments{
\item{data}{DataFrame that GeoLfit will use to determine a result.
Should be the output of \code{GeoDataRead}.}

\item{treatment_locations}{Vector of locations where the treatment was applied.}

\item{treatment_start_period}{Integer representing period where test started.}

\item{treatment_end_period}{Integer representing period where test finished.}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{Y_id}{Name of the outcome variable (String).}
}
\value{
A dataframe that holds the accumulated lift effect throughout the entire treatment period.
}
\description{
This method will calculate the cumulative lift with each passing day.
}
