% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cumulative_value.plot}
\alias{cumulative_value.plot}
\title{Plot the accumulated lift effect.}
\usage{
cumulative_value.plot(
  data,
  treatment_locations,
  treatment_start_period,
  treatment_end_period,
  location_id = "location",
  time_id = "time",
  Y_id = "Y",
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  title = "",
  subtitle = "",
  notes = "",
  ...
)
}
\arguments{
\item{data}{DataFrame that GeoLfit will use to determine a result.
Should be the output of \code{GeoDataRead}.}

\item{treatment_locations}{Vector of locations where the treatment was applied.}

\item{treatment_start_period}{Integer representing period where test started.}

\item{treatment_end_period}{Integer representing period where test finished.}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{Y_id}{Name of the outcome variable (String).}

\item{treatment_end_date}{Character that represents a date in year-month-day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{title}{Character for the title of the plot. NULL by default.}

\item{subtitle}{Character for the subtitle of the plot. NULL by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{...}{additional arguments}
}
\value{
A ggplot object that shows the accumulated lift per time period.
}
\description{
Plot the accumulated lift effect.
}
