% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{GeoLift}
\alias{GeoLift}
\title{GeoLift inference calculation.}
\usage{
GeoLift(
  Y_id = "Y",
  X = c(),
  data,
  locations,
  treatment_start_time,
  treatment_end_time
)
}
\arguments{
\item{Y_id}{Name of the outcome variable (String).}

\item{X}{List of names of covariates.}

\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{List of test locations.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}
}
\value{

}
\description{
\code{GeoLift} performs inference for a geo-test.
}
