% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{GeoLift}
\alias{GeoLift}
\title{GeoLift inference calculation.}
\usage{
GeoLift(
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  X = c(),
  data,
  locations,
  treatment_start_time,
  treatment_end_time
)
}
\arguments{
\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{X}{List of names of covariates.}

\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{List of test locations.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}
}
\value{
GeoLift object that contains:
         \itemize{
         \item{"results":}{ Generalized Augmented Sunthetic Controls results.}
         \item{"inference":}{ Data frame with inference statistics (ATT, Lift, p-value, and Confidence Interval.)}
         \item{"data":}{ Input data.}
         \item{"y_obs":}{ Observed outcome metric.}
         \item{"y_hat":}{ Counterfactual outcome metric.}
         \item{"ATT":}{ ATT estimate.}
         \item{"ATT_se":}{ Standrd Error of the ATT estimate.}
         \item{"TreatmentStart":}{ Time id of treatment start.}
         \item{"TreatmentEnd":}{ Time id of treatment end.}
         \item{"test_id":}{ List of names of the test locations.}
         \item{"incremental":}{ Incremental outcome units (Obersved - Counterfactual).}
         \item{"Y_id":}{ Name of the outcome variable.}
         }
}
\description{
\code{GeoLift} performs inference for a geo-test.
}
