% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_test_analysis.R, R/plots.R, R/imports.R
\docType{package}
\name{GeoLift}
\alias{GeoLift}
\alias{print.GeoLift}
\alias{summary.GeoLift}
\alias{print.summary.GeoLift}
\alias{plot.GeoLift}
\alias{GeoLift-package}
\title{GeoLift inference calculation.}
\usage{
GeoLift(
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  X = c(),
  data,
  locations,
  treatment_start_time,
  treatment_end_time,
  alpha = 0.1,
  model = "none",
  fixed_effects = TRUE,
  ConfidenceIntervals = FALSE,
  method = "conformal",
  grid_size = 250,
  stat_test = "Total"
)

\method{print}{GeoLift}(x, ...)

\method{summary}{GeoLift}(object, ...)

\method{print}{summary.GeoLift}(x, ...)

\method{plot}{GeoLift}(
  x,
  type = "Lift",
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  title = "",
  subtitle = "",
  notes = "",
  post_treatment_periods = 0,
  ...
)
}
\arguments{
\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{X}{List of names of covariates.}

\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{List of test locations.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{alpha}{Significance level. Set to 0.1 by default.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
\item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
\item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
for smaller panels (less than 40 locations and 100 time-stamps.))}
\item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
to improve fit for larger panels (more than 40 locations and 100
time-stamps. }
\item{"best:}{ Fits the model with the lowest Scaled L2 Imbalance.}
}}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}

\item{ConfidenceIntervals}{A logic flag indicating whether to estimate confidence intervals.
Set to FALSE by default.}

\item{method}{A string indicating the method used to calculate the
aggregate ATT Confidence Intervals.
\itemize{
\item{"conformal":}{ Conformal Inference. Defualt.}
\item{"jackknife+":}{ Jackknife+ (exclusively for stat_test = "Total").}
}}

\item{grid_size}{Number of grid points to use when inverting the hypothesis
test for Conformal Inference. Set to 250 by default.}

\item{stat_test}{A string indicating the test statistic.
\itemize{
\item{"Total":}{ The test statistic is the sum of all treatment effects, i.e. sum(abs(x)). Default.}
\item{"Negative":}{ One-sided test against positive effects i.e. -sum(x).
Recommended for Negative Lift tests.}
\item{"Positive":}{ One-sided test against negative effects i.e. sum(x).
Recommended for Positive Lift tests.}
}}

\item{x}{\code{GeoLift()}.}

\item{...}{additional arguments}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted. If type is set to "Incrementality",
daily incremental values are plotted.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{title}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{post_treatment_periods}{Number of post-treatment periods. Zero by default.}
}
\value{
GeoLift object that contains:
\itemize{
\item{"results":}{ Generalized Augmented Sunthetic Controls results.}
\item{"inference":}{ Data frame with inference statistics (ATT, Lift, p-value, and Confidence Interval.)}
\item{"data":}{ Input data.}
\item{"y_obs":}{ Observed outcome metric.}
\item{"y_hat":}{ Counterfactual outcome metric.}
\item{"ATT":}{ ATT estimate.}
\item{"ATT_se":}{ Standrd Error of the ATT estimate.}
\item{"TreatmentStart":}{ Time id of treatment start.}
\item{"TreatmentEnd":}{ Time id of treatment end.}
\item{"test_id":}{ List of names of the test locations.}
\item{"incremental":}{ Incremental outcome units (Obersved - Counterfactual).}
\item{"Y_id":}{ Name of the outcome variable.}
\item{"summary":}{ Model's Summary.}
\item{"ConfidenceIntervals":}{ Flag indicating whether CI will be included.}
\item{"lower_bound":}{ Lower confidence interval.}
\item{"upper_bound":}{ Upper confidence interval.}
}

GeoLift summary object that contains:
\itemize{
\item{"ATT":}{ ATT estimate.}
\item{"PercLift":}{ Lift estimate}
\item{"pvalue":}{ Experiment p-value.}
\item{"LowerCI":}{ Lower Confidence Interval.}
\item{"UpperCI":}{ Upper Confidence Interval.}
\item{"GlobalL2Imbalance":}{ Global L2 Imbalance.}
\item{"GlobalL2ImbalanceScaled":}{ Scaled Global L2 Imbalance.}
\item{"IndL2Imbalance":}{ Individual L2 Imbalance.}
\item{"IndL2ImbalanceScaled":}{ Scaled Individual L2 Imbalance.}
\item{"ATT":}{ ATT.}
\item{"start":}{ Treatment start.}
\item{"end":}{ Treatment end.}
\item{"type":}{ Single or Multiple test locations.}
\item{"Y_id":}{ Name of the outcome variable.}
\item{"incremental":}{ Incremental outcome units.}
\item{"bias":}{ Estimated corrected bias.}
\item{"weights":}{ Synthetic Control Weights.}
\item{"CI":}{ Flag indicating whether to include Confidence Intervals.}
\item{"alpha":}{ Significance level.}
\item{"lower":}{ Lower Bound Confidence Interval.}
\item{"upper":}{ Upper Bound Confidence Interval.}
}
}
\description{
\code{GeoLift} performs inference for a geo-test.

GeoLift is an end-to-end geo-experimental methodology based on
Synthetic Control Methods used to measure the true incremental effect (Lift)
of an ad campaign.
}
\author{
\strong{Maintainer}: Arturo Esquerra \email{aesquerra@fb.com}

Authors:
\itemize{
  \item Nicolas Cruces \email{nicocru@fb.com}
}

Other contributors:
\itemize{
  \item Kanishka Panchal \email{kpanchal@fb.com} [contributor]
  \item Michael Khalil \email{michaelkhalil@fb.com} [contributor]
}

}
