% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{GeoLift}
\alias{GeoLift}
\title{GeoLift inference calculation.}
\usage{
GeoLift(
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  X = c(),
  data,
  locations,
  treatment_start_time,
  treatment_end_time,
  alpha = 0.1,
  model = "none",
  fixed_effects = TRUE,
  ConfidenceIntervals = FALSE,
  stat_test = "Total",
  print = TRUE
)
}
\arguments{
\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{X}{List of names of covariates.}

\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{List of test locations.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{alpha}{Significance level. Set to 0.1 by default.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
         \item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
         \item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
                          for smaller panels (less than 40 locations and 100 time-stamps.))}
         \item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
                         to improve fit for larger panels (more than 40 locations and 100
                         time-stamps. }
         \item{"best:}{ Fits the model with the lowest Scaled L2 Imbalance.}
         }}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}

\item{ConfidenceIntervals}{A logic flag indicating whether to estimate confidence intervals
through jackknifing techniques.  Set to FALSE by default.}

\item{stat_test}{A string indicating the test statistic.
\itemize{
         \item{"Total":}{ The test statistic is the sum of all treatment effects, i.e. sum(abs(x)). Default.}
         \item{"Negative":}{ One-sided test against positive effects i.e. -sum(x).
         Recommended for Negative Lift tests.}
         \item{"Positive":}{ One-sided test against negative effects i.e. sum(x).
         Recommended for Positive Lift tests.}
}}

\item{print}{A logic flag indicating whether to print the results or not.
Set to TRUE by default.}
}
\value{
GeoLift object that contains:
         \itemize{
         \item{"results":}{ Generalized Augmented Sunthetic Controls results.}
         \item{"inference":}{ Data frame with inference statistics (ATT, Lift, p-value, and Confidence Interval.)}
         \item{"data":}{ Input data.}
         \item{"y_obs":}{ Observed outcome metric.}
         \item{"y_hat":}{ Counterfactual outcome metric.}
         \item{"ATT":}{ ATT estimate.}
         \item{"ATT_se":}{ Standrd Error of the ATT estimate.}
         \item{"TreatmentStart":}{ Time id of treatment start.}
         \item{"TreatmentEnd":}{ Time id of treatment end.}
         \item{"test_id":}{ List of names of the test locations.}
         \item{"incremental":}{ Incremental outcome units (Obersved - Counterfactual).}
         \item{"Y_id":}{ Name of the outcome variable.}
         \item{"summary":}{ Model's Summary.}
         \item{"ConfidenceIntervals":}{ Flag indicating whether CI will be included.}
         \item{"lower_bound":}{ Lower confidence interval.}
         \item{"upper_bound":}{ Upper confidence interval.}
         }
}
\description{
\code{GeoLift} performs inference for a geo-test.
}
