% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R, R/plots.R
\name{GeoLiftMultiCell}
\alias{GeoLiftMultiCell}
\alias{print.GeoLiftMultiCell}
\alias{summary.GeoLiftMultiCell}
\alias{plot.GeoLiftMultiCell}
\title{GeoLiftMultiCell inference calculation.}
\usage{
GeoLiftMultiCell(
  Y_id = "Y",
  time_id = "time",
  location_id = "location",
  X = c(),
  data,
  locations,
  treatment_start_time,
  treatment_end_time,
  alpha = 0.1,
  model = "none",
  fixed_effects = TRUE,
  ConfidenceIntervals = TRUE,
  method = "conformal",
  grid_size = 250,
  stat_test = "Total",
  winner_declaration = TRUE,
  geolift_type = "standard"
)

\method{print}{GeoLiftMultiCell}(x, ...)

\method{summary}{GeoLiftMultiCell}(object, table = FALSE, ...)

\method{plot}{GeoLiftMultiCell}(
  x,
  type = "Lift",
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  title = "",
  post_treatment_periods = 0,
  stacked = TRUE,
  ...
)
}
\arguments{
\item{Y_id}{Name of the outcome variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{X}{List of names of covariates.}

\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{A list of lists of test markets per cell. The recommended layout is
\code{list(cell_1 = list("locA"), cell2 = list("locB"), cell3 = list("locC"),...)}.}

\item{treatment_start_time}{Time index of the start of the treatment.}

\item{treatment_end_time}{Time index of the end of the treatment.}

\item{alpha}{Significance level. Set to 0.1 by default.}

\item{model}{A string indicating the outcome model used to augment the Augmented
Synthetic Control Method. Augmentation through a prognostic function can improve
fit and reduce L2 imbalance metrics.
\itemize{
\item{"None":}{ ASCM is not augmented by a prognostic function. Defualt.}
\item{"Ridge":}{ Augments with a Ridge regression. Recommended to improve fit
for smaller panels (less than 40 locations and 100 time-stamps.))}
\item{"GSYN":}{ Augments with a Generalized Synthetic Control Method. Recommended
to improve fit for larger panels (more than 40 locations and 100
time-stamps. }
\item{"best:}{ Fits the model with the lowest Scaled L2 Imbalance.}
}}

\item{fixed_effects}{A logic flag indicating whether to include unit fixed
effects in the model. Set to TRUE by default.}

\item{ConfidenceIntervals}{A logic flag indicating whether to estimate confidence intervals.
Set to FALSE by default.}

\item{method}{A string indicating the method used to calculate the
aggregate ATT Confidence Intervals.
\itemize{
\item{"conformal":}{ Conformal Inference. Defualt.}
\item{"jackknife+":}{ Jackknife+ (exclusively for stat_test = "Total").}
}}

\item{grid_size}{Number of grid points to use when inverting the hypothesis
test for Conformal Inference. Set to 250 by default.}

\item{stat_test}{A string indicating the test statistic.
\itemize{
\item{"Total":}{ The test statistic is the sum of all treatment effects, i.e. sum(abs(x)). Default.}
\item{"Negative":}{ One-sided test against positive effects i.e. -sum(x).
Recommended for Negative Lift tests.}
\item{"Positive":}{ One-sided test against negative effects i.e. sum(x).
Recommended for Positive Lift tests.}
}}

\item{winner_declaration}{Logic flag indicating whether to compute a winner cell analysis.
If set to TRUE (default), both pairwise and total statistical significance tests will be
performed.}

\item{geolift_type}{String that specifies the type of GeoLift test to be performed:
\itemize{
\item{"standard":}{ Standard GeoLift test where ads are shown to test regions. Defualt.}
\item{"inverse":}{ Inverse or Negative GeoLift test where the test group is holded-out
from the treatment.}
}}

\item{x}{\code{GeoLiftMultiCell()}}

\item{...}{additional arguments}

\item{object}{\code{GeoLiftMultiCell()}}

\item{table}{Logic flag indicating whether to plot only a table
summarizing the results or the entire verbose output of each cell's
GeoLift object summary. FALSE by default.}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted. If type is set to "Incrementality",
daily incremental values are plotted.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{title}{String for the title of the plot. Empty by default.}

\item{post_treatment_periods}{Number of post-treatment periods. Zero by default.}

\item{stacked}{Logic flag indicating whether to stack all the Multi-Cell plots
together vertically or to output each one of them separately. Set to TRUE by
default.}
}
\value{
\code{GeoLiftMultiCell} object that contains:
\itemize{
\item{"results":}{ List of \code{GeoLift} objects (one per cell).}
\item{"pairwise":}{ Table of pairwise comparisons and pairwise winner declarations.)}
\item{"Winner":}{ List of winner locations for the total comparison (if any).}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{GeoLiftMultiCell} performs inference for a Multi-Cell GeoLift.
}
