% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{GeoLiftPower}
\alias{GeoLiftPower}
\title{Power Calculation for GeoLift for known test locations.}
\usage{
GeoLiftPower(
  data,
  locations,
  effect_size = seq(0, 10000, 50)/1000,
  treatment_periods,
  horizon = 30,
  cpic = 0,
  X = c(),
  Y_id = "Y"
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locations}{A semi-colon separated list of test geo locations.}

\item{effect_size}{A vector of effect sizes to test by default a
sequence between 0 - 15 percent: seq(0,10000,50)/1000.}

\item{treatment_periods}{Expected length of the test. A vector of
possible lengths can be entered for multiple options.}

\item{horizon}{Lookback window for the power analysis. For daily
data 30 days are recommended and 8-12 weeks for monthly data.
The default window is 30.}

\item{cpic}{Cost Per Incremental Conversion for estimated test
minimum budget. The default value is 0, in which case no investment
estimation will be provided.}

\item{X}{List of names of covariates. No covariates are used
by default.}

\item{Y}{Name of the outcome variable. "Y" by default.}
}
\value{
GeoLiftPower object that contains:
     \itemize{
         \item{"location":}{ Test units of the simulation}
         \item{"pvalue":}{ P Value for each simulation}
         \item{"duration":}{ Duration of the simulation}
         \item{"effect_size":}{ Effect Size used for the simulation}
         \item{"treatment_start":}{ Treatment start time for the simulation}
         \item{"investment":}{ Estimated Investment}
     }
}
\description{
This function runs power calculations for input historical geo data
for a pre-determined set of test locations.
}
