% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{GeoLiftPower.search}
\alias{GeoLiftPower.search}
\title{Power calculations for unknown test market locations, number of
test markets, and test duration.}
\usage{
GeoLiftPower.search(
  data,
  treatment_periods,
  N = 1,
  horizon = 30,
  X = c(),
  Y_id = "Units",
  location_id = "location",
  time_id = "time",
  top_results = 5
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{treatment_periods}{List of treatment periods to calculate power for.}

\item{N}{List of number of test markets to calculate power for.}

\item{horizon}{Lookback window for the power analysis. For daily
data 30 days are recommended and 8-12 weeks for monthly data.
The default window is 30.}

\item{X}{List of names of covariates.}

\item{Y_id}{Name of the outcome variable (String).}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{top_results}{Number of results to display}
}
\value{
Data frame with the ordered list of best locations and their
average power.
}
\description{
\code{GeoLiftPower.search} provides power calculations for unknown
test markets, number of test locations, and test duration.
}
