% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_data.R
\name{is_location_in_cluster}
\alias{is_location_in_cluster}
\title{Define if location is in cluster, via lat and long.}
\usage{
is_location_in_cluster(location_points, cluster_polygon)
}
\description{
Define if location is in cluster, via lat and long.
}
