% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{Lift.plot}
\alias{Lift.plot}
\title{Aggregate Lift plot function for GeoLift.}
\usage{
Lift.plot(GeoLift, outcome, ...)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{outcome}{Name of the outcome variable. By default "Units".}
}
\value{
Aggregate Lift plot.
}
\description{
Aggregate Lift plot function for GeoLift.
}
