% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Lift.plot}
\alias{Lift.plot}
\title{Aggregate Lift plot function for GeoLift.}
\usage{
Lift.plot(
  GeoLift,
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  title = "",
  subtitle = "",
  notes = "",
  post_treatment_periods = 0,
  ...
)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{title}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{post_treatment_periods}{Number of post-treatment periods. Zero by default.}

\item{...}{additional arguments}
}
\value{
Aggregate Lift plot.
}
\description{
Aggregate Lift plot function for GeoLift.
}
