% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{Lift.plot}
\alias{Lift.plot}
\title{Aggregate Lift plot function for GeoLift.}
\usage{
Lift.plot(
  GeoLift,
  outcome = "Units",
  main = "",
  subtitle = "",
  notes = "",
  ...
)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{outcome}{String for the name of the outcome variable. By default "Units".}

\item{main}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{...}{additional arguments}
}
\value{
Aggregate Lift plot.
}
\description{
Aggregate Lift plot function for GeoLift.
}
