% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_data.R
\name{location_to_cluster_matching}
\alias{location_to_cluster_matching}
\title{Match all locations to the cluster they belong.}
\usage{
location_to_cluster_matching(
  location_points,
  clusters,
  longitude_col_name = "Longitude",
  latitude_col_name = "Latitude",
  verbose = FALSE
)
}
\arguments{
\item{location_points}{data.frame object that holds the location name,
latitude and longitude for all locations.}

\item{clusters}{data.frame object with all polygons.  Each row represents a
different cluster with an sf::POLYGON object that has the points that are joined
to form that cluster.}

\item{longitude_col_name}{name of the longitude column in the location_points
data.frame.}

\item{latitude_col_name}{name of the latitude column in the location_points
data.frame.}

\item{verbose}{logical flag indicating whether amount of clusters processed
should be counted.}
}
\value{
A data frame holding a match between each location and all clusters.
}
\description{
Use clusters defined by how users move between cities to decide how cities
should be aggregated to reduce contamination.
}
