% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{MarketCorrelations}
\alias{MarketCorrelations}
\title{Get within market correlations for all locations.}
\usage{
MarketCorrelations(data)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}
}
\value{
A Dataframe where each column represents the closest market to the
market in the first column, ordering them by their correlation factor.
}
\description{
\code{get_market_correlations} calculates a notion of similarity between markets
to help inform the combinations of treatments.
}
