% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{MarketSelection}
\alias{MarketSelection}
\title{Market selection tool.}
\usage{
MarketSelection(
  data,
  location_id = "location",
  time_id = "time",
  Y_id = "Y",
  dtw = 0
)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{location_id}{Name of the location variable (String).}

\item{time_id}{Name of the time variable (String).}

\item{Y_id}{Name of the outcome variable (String).}

\item{dtw}{Emphasis on Dynamic Time Warping (DTW), dtw = 1 focuses exclusively
on this metric while dtw = 0 (default) relies on correlations only.}
}
\value{
Matrix of the best markets. The second to last columns show
the best to worst market matches for the location in the first
column.
}
\description{
\code{MarketSelection} helps calculate the best markets based
on Dynamic Time Warping between the locations' time-series.
}
