% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{plot.GeoLift}
\alias{plot.GeoLift}
\title{Plot for GeoLift.}
\usage{
\method{plot}{GeoLift}(
  GeoLift,
  type = "Lift",
  outcome = "Units",
  conf.level = 0.9,
  test_locs = FALSE,
  ...
)
}
\arguments{
\item{GeoLift}{GeoLift object.}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted.}

\item{outcome}{Name of the outcome variable. By default "Units".}

\item{conf.level}{Confidence level. By defaul 0.9.}

\item{test_locs}{Plot the average results by default. If set to
TRUE, the ATT by test location is plotted.}
}
\value{
GeoLift plot.
}
\description{
Plot for GeoLift objects.
}
