% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{plot.GeoLift}
\alias{plot.GeoLift}
\title{Plot for GeoLift.}
\usage{
\method{plot}{GeoLift}(
  x,
  type = "Lift",
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  title = "",
  subtitle = "",
  notes = "",
  ...
)
}
\arguments{
\item{x}{GeoLift object.}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted. If type is set to "Incrementality",
daily incremental values are plotted.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{title}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{...}{additional arguments}
}
\value{
GeoLift plot.
}
\description{
Plot for GeoLift objects.
}
