% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{plot.GeoLift}
\alias{plot.GeoLift}
\title{Plot for GeoLift.}
\usage{
\method{plot}{GeoLift}(
  x,
  type = "Lift",
  outcome = "Units",
  main = "",
  subtitle = "",
  notes = "",
  conf.level = 0.9,
  test_locs = FALSE,
  ...
)
}
\arguments{
\item{x}{GeoLift object.}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted.}

\item{outcome}{String name of the outcome variable. By default "Units".}

\item{main}{String for the title of the plot. Empty by default.}

\item{subtitle}{String for the subtitle of the plot. Empty by default.}

\item{notes}{String to add notes to the plot. Empty by default.}

\item{conf.level}{Confidence level. By defaul 0.9.}

\item{test_locs}{Plot the average results by default. If set to
TRUE, the ATT by test location is plotted.}

\item{...}{additional arguments}
}
\value{
GeoLift plot.
}
\description{
Plot for GeoLift objects.
}
