% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{plot.GeoLiftPower}
\alias{plot.GeoLiftPower}
\title{Plotting function for GeoLiftPower.}
\usage{
\method{plot}{GeoLiftPower}(x, power = 0.8, conf.level = 0.9, table = TRUE, ...)
}
\arguments{
\item{x}{GeoLiftPower object.}

\item{power}{Power level. By default 0.8.}

\item{conf.level}{Confidence level. By default 0.9.}

\item{table}{Plot the table of power estimates. TRUE by default.}

\item{...}{additional arguments}
}
\value{
GeoLiftPower plot.
}
\description{
Plotting functoin for \code{GeoLiftPower}.
}
