% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{plot.GeoLiftPower}
\alias{plot.GeoLiftPower}
\title{Plotting function for GeoLiftPower.}
\usage{
\method{plot}{GeoLiftPower}(x, power = 0.8, table = TRUE, actual_values = FALSE, ...)
}
\arguments{
\item{x}{GeoLiftPower object.}

\item{power}{Power level. By default 0.8.}

\item{table}{Plot the table of power estimates. TRUE by default.}

\item{actual_values}{Logic flag indicating whether to include in the plot
the actual values in addition to the smoothed values.}

\item{...}{additional arguments}
}
\value{
GeoLiftPower plot.
}
\description{
Plotting function for \code{GeoLiftPower}. The function smooths the power curve
for ease of interpretation.
}
