% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{plot.MultiCellMarketSelection}
\alias{plot.MultiCellMarketSelection}
\title{Plotting function for MultiCellMarketSelection}
\usage{
\method{plot}{MultiCellMarketSelection}(
  x,
  test_markets = list(),
  type = "Lift",
  treatment_end_date = NULL,
  frequency = "daily",
  plot_start_date = NULL,
  post_treatment_periods = 0,
  title = "",
  stacked = TRUE,
  ...
)
}
\arguments{
\item{x}{A MultiCellMarketSelection object.}

\item{test_markets}{List of market IDs per cell. The list must contain exactly
k numeric values corresponding to the power analysis. The recommended layout is
\code{list(cell_1 = 1, cell2 = 1, cell3 = 1,...)}.}

\item{type}{Type of plot. By default "Lift" which plots the
incrementality on the outcome variable. If type is set to "ATT",
the average ATT is plotted. If type is set to "Incrementality",
daily incremental values are plotted.}

\item{treatment_end_date}{Character that represents a date in year-month=day format.}

\item{frequency}{Character that represents periodicity of time stamps. Can be either
weekly or daily. Defaults to daily.}

\item{plot_start_date}{Character that represents initial date of plot in year-month-day format.}

\item{post_treatment_periods}{Number of post-treatment periods. Zero by default.}

\item{title}{String for the title of the plot. Empty by default.}

\item{stacked}{Logic flag indicating whether to stack all the Multi-Cell plots
together vertically or to output each one of them separately. Set to TRUE by
default.}

\item{...}{additional arguments}
}
\value{
MultiCellMarketSelection plot.
}
\description{
Plotting function for \code{MultiCellMarketSelection}. This function plots the
latest possible test given the data and duration as well as the power curve
across historical simulations for a given market_id and cell_id.
}
