% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCorrels}
\alias{plotCorrels}
\title{Plotting function for historical correlations between the test market's
and total markets KPI.}
\usage{
plotCorrels(data, locs = c(), scaled = TRUE, control_corr = TRUE)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit. It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{locs}{List of markets to use in the calculation of the correlations.}

\item{scaled}{A logic flag indicating whether to plot the scaled values of
the KPI metric (standardized by the largest historical value). Set to FALSE
by default to plot the observed KPI levels.}

\item{control_corr}{A logic flag indicating how correlations will be computed.
If set to TRUE (default), correlations between the test group and pool of
controls are computed. If set to FALSE, correlations between test units and
the locations in the entire data set are calculated.}
}
\value{
A plot of the historical values of the test market and the aggregation of
all control markets.
}
\description{
\code{plotCorrels} takes a data frame used for GeoLift to generate
a plot that show historical similarities in KPI levels between the
test markets and the aggregation of all locations (total).
}
