% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLift.R
\name{pvalueCalc}
\alias{pvalueCalc}
\title{Calculate p-value for GeoLift.}
\usage{
pvalueCalc(data, sim, max_time, tp, es, locations, cpic, X)
}
\arguments{
\item{data}{A data.frame containing the historical conversions by
geographic unit, It requires a "locations" column with the geo name,
a "Y" column with the outcome data (units), a time column with the indicator
of the time period (starting at 1), and covariates.}

\item{sim}{Time simulation index.}

\item{max_time}{Treatment end index.}

\item{tp}{Time period index.}

\item{es}{Effect Size.}

\item{locations}{List of test locations.}

\item{cpic}{Cost Per Incremental Conversion.}

\item{X}{List of names of covariates.}
}
\value{
List that contains:
         \itemize{
         \item{"location":}{ Test locations.}
         \item{"pvalue":}{ P Value.}
         \item{"tp":}{ Time period index.}
         \item{"es":}{ Effect Size used for the simulation.}
         \item{"treatment_start_time":}{ Treatment start time for the simulation}
         \item{"investment":}{ Estimated Investment}
        }
}
\description{
\code{pvalueCalc} calculates the p-value for a GeoLift object.
}
