% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_data.R
\name{run_cluster_matching}
\alias{run_cluster_matching}
\title{Match all locations to the cluster they belong.}
\usage{
run_cluster_matching(
  data,
  location_id = "location",
  X = c(),
  country_name = NULL,
  longitude_col_name = "longitude",
  latitude_col_name = "latitude",
  find_location_lat_long = FALSE,
  gmaps_key = NULL
)
}
\arguments{
\item{data}{data.frame object that holds the location name,
latitude and longitude for all locations.}

\item{location_id}{Name of the location variable (String).}

\item{X}{List of names of the covariates.}

\item{country_name}{name of country where the locations come from.}

\item{longitude_col_name}{name of the longitude column in the \code{data}
data.frame.}

\item{latitude_col_name}{name of the latitude column in the \code{data}
data.frame.}

\item{find_location_lat_long}{logical flag indicating whether Google Maps API
should be used to find location's latitude and longitude.}

\item{gmaps_key}{string containing your Google Maps API key.}
}
\value{
A data frame holding a match between each location and all clusters.
}
\description{
Use clusters defined by how users move between cities to decide how cities
should be aggregated to reduce contamination.
}
