% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{summary.GeoLiftMultiCell}
\alias{summary.GeoLiftMultiCell}
\title{Summary method for GeoLiftMultiCell}
\usage{
\method{summary}{GeoLiftMultiCell}(object, ...)
}
\arguments{
\item{object}{GeoLiftMultiCell object.}

\item{...}{Optional arguments}
}
\description{
GeoLiftMultiCell summary output with additional information about the
test.
}
